backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
scipy
pint
openpyxl
pandas
sqlalchemy
pytest
pytest-cov
python-multipart
moorpy
reportlab
httpx
pypdf
slowapi
python-dotenv
